Moved ApiRequestBuilder not null input stream test to own class
